+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Added missing `Nanook::WalletAccount#block_count` delegate.
+- Added `Nanook::Node#confirmation_quorum`.
 
 ### Changed